+++ compile-and-test.sh
@@ -6,6 +6,12 @@ set -e
 
 INFLUXDB_VERSIONS="1.3 1.2 1.1"
 
+JAVA_VERSIONS="3-jdk-8 3-jdk-9"
+
+
+for java_version in ${JAVA_VERSIONS}
+do
+ echo "Run tests with maven:${java_version}"
 for version in ${INFLUXDB_VERSIONS}
 do
   echo "Tesing againts influxdb ${version}"
@@ -26,7 +32,8 @@ do
         --workdir /usr/src/mymaven \
         --link=influxdb \
         --env INFLUXDB_IP=influxdb \
-         maven:alpine mvn clean install
+         maven:${java_version} mvn clean install
 
   docker kill influxdb || true
+done
 done